fix(chat): route directive MCP tools through chat.api import boundary
